requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.29.0
pymssql==2.2.11
python-dotenv==1.0.1
PyJWT==2.8.0
pydantic==2.7.4
pydantic-settings==2.4.0
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
